+++ src/main/java/cart/dummydata/MemberInitializer.java
@@ -1,4 +1,4 @@
-package cart.dao.dummyData;
+package cart.dummydata;
 
 import cart.dao.MemberDao;
 import cart.domain.entity.Member;
