+++ src/test/java/cart/dao/JdbcCartItemDaoTest.java
@@ -1,7 +1,7 @@
 package cart.dao;
 
-import cart.dao.dummyData.MemberInitializer;
-import cart.dao.dummyData.ProductInitializer;
+import cart.dummydata.MemberInitializer;
+import cart.dummydata.ProductInitializer;
 import cart.domain.entity.CartItem;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
